+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Dev: Improve Twitch.tv clip tests. (#283)
 - Dev: Improve YouTube tests. (#284)
 - Dev: Resolver Check now returns a context. (#287)
+- Dev: Improve Wikipedia tests. (#286)
 
 ## 1.2.3